functions/version1/framer/requirements.txt
opencv-python-headless==4.5.4.60
numpy==1.21.4
minio==7.1.1
Pillow==8.4.0